requirements.txt
uvicorn==0.34.0
fastapi==0.115.11
sqlmodel==0.0.24
pydantic-settings==2.8.1
pydantic[email]==2.10.6
psycopg[binary]==3.2.5
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
fastapi-cache2==0.2.2
redis==5.2.1
pytest==8.3.5
httpx==0.28.1